requirements.txt
mypy
torch
ffmpeg-python
git+https://github.com/JulianKropp/stream_pipeline

python-dotenv
cryptography
flask
prometheus_client
numpy
faster-whisper @ https://github.com/SYSTRAN/faster-whisper/archive/refs/heads/master.tar.gz
